Backend/ATProject/requirements.txt
python-dotenv==1.0.1
psycopg2-binary
django
djangorestframework
django-cors-headers